# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Gaze Engine API
  version: 3.0.0
  description: Privacy-first local video search engine API
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "48100"
security:
  - bearerAuth: []
paths:
...
  /models:
    get:
      operationId: listModels
      summary: List available models and download status
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
    post:
      operationId: downloadModel
      summary: Start model download
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadModelRequest"
      responses:
        "200":
          description: Download started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadModelResponse"
  /models/{model_name}/progress:
    get:
      operationId: getModelDownloadProgress
      summary: Get download progress for a specific model
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
            enum: [whisper-base, openclip-vit-b-32, ssdlite320-mobilenet-v3]
      responses:
        "200":
          description: Download progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [complete, downloading, pending, error]
                  progress:
                    type: number
                    minimum: 0
                    maximum: 1
                  error:
                    type: string
        "404":
          description: Model not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ModelInfo:
      type: object
      required:
        - name
        - downloaded
        - size_bytes
      properties:
        name:
          type: string
        downloaded:
          type: boolean
        size_bytes:
          type: integer
        download_progress:
          type: number
          minimum: 0
          maximum: 1
    ModelsResponse:
      type: object
      required:
        - models
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelInfo"
    DownloadModelRequest:
      type: object
      required:
        - model
      properties:
        model:
          type: string
          enum: [whisper-base, openclip-vit-b-32, ssdlite320-mobilenet-v3]
    DownloadModelResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [started, already_downloaded, downloading]
...